requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlmodel>=0.0.22,<1.0
pydantic-settings>=2.6,<3.0
python-multipart>=0.0.18,<1.0
reportlab>=4.2,<5.0
python-docx>=1.1,<2.0
psycopg[binary]>=3.2,<4.0
